refactor: :art: Airline Manager unique identifierNumber custom validation
